selfhost/stage07/include/stddef.h: track host pointer width for size_t
The stub had \`typedef unsigned int size_t\` (always 4 bytes) which
hardcoded the SLOW-32 native size.  On a 64-bit host (cc-x64, cc-a64
— both set ty_ptr_size=8 in main, which makes the parser default
size_t = unsigned long long) this typedef *overrode* the parser
default to a 4-byte int.

Visible symptom in dbt-a64: every emit_ctx_t and translate_ctx_t had
their size_t-sized fields (capacity, offset, jmp_patch_offset) packed
as 4 bytes, shifting all later struct members.  ctx->cache landed at
offset 56 instead of 64 in cc-a64's view, vs. the runtime layout
mmio_ring/cpu_state-style code expected.  Codex's commit aed3584
made the prologue parallel-copy correct; this completes the fix on
the type side.

Use \`unsigned long\` and \`long\`: cc-a64/cc-x64's parser maps these
through target_long_ty(), which returns TY_LLONG (8 bytes) when
ty_ptr_size=8 and TY_INT (4 bytes) when ty_ptr_size=4 — so the same
header works for SLOW-32 native, cc-x64, and cc-a64 without an
ifdef.

Result on dbt-a64: emit_ctx_t now has the expected 44-byte layout
(8+8+8+4+4+4+8 = 44), and translate_ctx_t->cache is at offset 64.
All 30 cc-a64 tests still pass; s32fast-hir builds.  dbt-a64 still
SIGSEGVs at a different point — the next bug is in this same area
but distinct from the size_t mismatch.

+++ selfhost/stage07/include/stddef.h
@@ -1,9 +1,14 @@
-/* stddef.h -- s12cc-compatible stub */
+/* stddef.h -- s12cc-compatible stub.  size_t / ptrdiff_t track the
+ * pointer width: `unsigned long` is 8 bytes on a 64-bit host (cc-x64,
+ * cc-a64 — both set ty_ptr_size=8) and 4 bytes on SLOW-32 native
+ * (target_long_ty() returns TY_INT when ty_ptr_size=4).  Hardcoding
+ * `unsigned int` here truncates pointer-sized fields (e.g. emit_ctx_t's
+ * size_t members) on 64-bit hosts and silently corrupts struct layouts. */
 #ifndef _STDDEF_H
 #define _STDDEF_H
 
-typedef unsigned int size_t;
-typedef int ptrdiff_t;
+typedef unsigned long size_t;
+typedef long ptrdiff_t;
 
 #define NULL 0
